+++ src/main/java/bowling/dto/PlayerResult.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.dto;
 
 import java.util.List;
